added admin views guard

+++ routes/web.php
@@ -9,34 +9,36 @@ use App\Http\Controllers\RoomController;
 use App\Http\Controllers\AuthController;
 
 Route::get('/',[HomeController::class,"index"]);
-Route::get('/hotels', [HotelsController::class, 'index']);
-Route::get('/hotels/edit/{id}',[HotelsController::class,"editView"]);
-Route::post('/hotels/update/{id}',[HotelsController::class,"update"]);
-Route::get('/hotels/create', [HotelsController::class, 'createView']);
-Route::post('/hotels/add-to-database', [HotelsController::class, 'addToDatabase']);
 
-Route::get('/amenities', [AmenityController::class, 'index']);
-Route::get('/amenities/edit/{id}',[AmenityController::class,"getAmenityView"]);
-Route::get('/amenities/create', [AmenityController::class, 'getCreateView']);
-Route::post('/amenities/add-to-database', [AmenityController::class, 'create']);
-Route::post('/amenities/update/{id}',[AmenityController::class,"update"]);
-Route::delete('/amenities/delete/{id}',[AmenityController::class,"delete"]);
+Route::get('/hotels', [HotelsController::class, 'index'])->middleware('admin');
+Route::get('/hotels/edit/{id}',[HotelsController::class,"editView"])->middleware('admin');
+Route::post('/hotels/update/{id}',[HotelsController::class,"update"])->middleware('admin');
+Route::get('/hotels/create', [HotelsController::class, 'createView'])->middleware('admin');
+Route::post('/hotels/add-to-database', [HotelsController::class, 'addToDatabase'])->middleware('admin') ;
+Route::delete('/hotels/delete/{id}', [HotelsController::class, 'delete'])->middleware('admin');
 
-Route::get('/rooms', [RoomController::class, 'index']);
-Route::get('/rooms/edit/{id}',[RoomController::class,"editView"]);
-Route::get('/rooms/create', [RoomController::class, 'createView']);
-Route::post('/rooms/add-to-database', [RoomController::class, 'addToDatabase']);
-Route::post('/rooms/update/{id}',[RoomController::class,"updateRoom"]);
-Route::get('/rooms/add-amenity/{id}',[RoomController::class, 'addAmenity']);
-Route::post('/rooms/add-amenity/{id}',[RoomController::class, 'addAmenityToDatabase']);
-// Route::delete('/rooms/delete/{id}',[RoomController::class,"delete"]);
+Route::get('/amenities', [AmenityController::class, 'index'])->middleware('admin');
+Route::get('/amenities/edit/{id}',[AmenityController::class,"getAmenityView"])->middleware('admin');
+Route::get('/amenities/create', [AmenityController::class, 'getCreateView'])->middleware('admin');
+Route::post('/amenities/add-to-database', [AmenityController::class, 'create'])->middleware('admin')    ;
+Route::post('/amenities/update/{id}',[AmenityController::class,"update"])->middleware('admin');
+Route::delete('/amenities/delete/{id}',[AmenityController::class,"delete"])->middleware('admin');
 
-Route::get('/reservations', [ReservationController::class, 'index']);
-Route::get('/reservations/edit/{id}',[ReservationController::class,"editView"]);
-Route::get('/reservations/create', [ReservationController::class, 'createView']);
-Route::post('/reservations/add-to-database', [ReservationController::class, 'addToDatabase']);
-Route::post('/reservations/update/{id}',[ReservationController::class,"updateReservation"]);
-Route::post('/reservations/delete/{id}',[ReservationController::class,"deleteReservation"]);
+Route::get('/rooms', [RoomController::class, 'index'])->middleware('admin');
+Route::get('/rooms/edit/{id}',[RoomController::class,"editView"])->middleware('admin');
+Route::get('/rooms/create', [RoomController::class, 'createView'])->middleware('admin');
+Route::post('/rooms/add-to-database', [RoomController::class, 'addToDatabase'])->middleware('admin');
+Route::post('/rooms/update/{id}',[RoomController::class,"updateRoom"])->middleware('admin');
+Route::get('/rooms/add-amenity/{id}',[RoomController::class, 'addAmenity'])->middleware('admin');
+Route::post('/rooms/add-amenity/{id}',[RoomController::class, 'addAmenityToDatabase'])->middleware('admin');
+Route::delete('/rooms/delete/{id}', [RoomController::class, 'delete'])->middleware('admin');
+
+Route::get('/reservations', [ReservationController::class, 'index'])->middleware('admin');
+Route::get('/reservations/edit/{id}',[ReservationController::class,"editView"])->middleware('admin');
+Route::get('/reservations/create', [ReservationController::class, 'createView'])->middleware('admin');
+Route::post('/reservations/add-to-database', [ReservationController::class, 'addToDatabase'])->middleware('admin');
+Route::post('/reservations/update/{id}',[ReservationController::class,"updateReservation"])->middleware('admin');
+Route::post('/reservations/delete/{id}',[ReservationController::class,"deleteReservation"])->middleware('admin')    ;
 
 Route::get('/login', [AuthController::class, 'loginView']);
 Route::post('/login', [AuthController::class, 'login']);
